requirements.txt
# requirements.txt
pandas==2.2.2
python-dotenv==1.0.1
qdrant-client==1.12.1
transformers==4.45.0
sentence-transformers==3.0.1
torch>=2.1.0
streamlit==1.38.0
openai==1.51.0
numpy==1.26.4
uvicorn==0.30.0